Fix the console problem and refactor the routes
