+++ client/src/App.jsx
@@ -1,55 +1,15 @@
 import React from 'react';
-import { createBrowserRouter, RouterProvider, Outlet } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 import './index.css';
 import Navbar from './components/navbar/Navbar';
 import Footer from './components/footer/Footer';
-import Home from './pages/home/Home';
-import Gig from './pages/gig/Gig';
-import Gigs from './pages/gigs/Gigs';
-import Orders from './pages/orders/Orders';
-import MyGigs from './pages/myGigs/MyGigs';
-import AddGig from './pages/addgig/AddGig';
-import Messages from './pages/messages/Messages';
-import Message from './pages/message/Message';
-import Login from './pages/login/Login';
-import Register from './pages/register/Register';
-import Success from './pages/success/Success';
-import Pay from './pages/pay/Pay';
 
 function App() {
-  const Layout = () => {
-    return (
-      <div className="flex flex-col min-h-screen">
-        <Navbar />
-        <Outlet />
-        <Footer />
-      </div>
-    );
-  };
-
-  const router = createBrowserRouter([
-    {
-      path: '/',
-      element: <Layout />,
-      children: [
-        { path: '/', element: <Home /> },
-        { path: '/gig/:id', element: <Gig /> },
-        { path: '/gigs', element: <Gigs /> },
-        { path: '/orders', element: <Orders /> },
-        { path: '/mygigs', element: <MyGigs /> },
-        { path: '/addgig', element: <AddGig /> },
-        { path: '/messages', element: <Messages /> },
-        { path: '/message/:id', element: <Message /> },
-        { path: '/login', element: <Login /> },
-        { path: '/register', element: <Register /> },
-        { path: '/pay/:id', element: <Pay /> },
-        { path: '/success', element: <Success /> },
-      ],
-    },
-  ]);
   return (
-    <div>
-      <RouterProvider router={router}></RouterProvider>
+    <div className="flex flex-col min-h-screen">
+      <Navbar />
+      <Outlet />
+      <Footer />
     </div>
   );
 }
